+++ Moobeez/Classes/Moobeez/Views/EpisodeCell.m
@@ -41,7 +41,7 @@
     
     [self.posterImageView loadImageWithPath:episode.posterPath andWidth:92 completion:^(BOOL didLoadImage) {}];
     self.titleLabel.text = episode.name;
-    self.episodeLabel.text = StringInteger(episode.episodeNumber);
+    self.episodeLabel.text = StringInteger((long)episode.episodeNumber);
 
     if (episode.description) {
         self.detailsLabel.text = episode.description;
@@ -83,7 +83,7 @@
         if ([[Database sharedDatabase] watch:NO episode:self.teebeeEpisode forTeebee:self.teebee]) {
             self.teebee.watchedEpisodesCount--;
             self.teebee.notWatchedEpisodesCount++;
-            self.teebee.seasons[StringInteger(self.teebeeEpisode.seasonNumber)] = @([self.teebee.seasons[StringInteger(self.teebeeEpisode.seasonNumber)] integerValue] - 1);
+            self.teebee.seasons[StringInteger(self.teebeeEpisode.seasonNumber)] = @([self.teebee.seasons[StringInteger((long)self.teebeeEpisode.seasonNumber)] integerValue] - 1);
             
             self.watchButton.selected = NO;
             
@@ -101,7 +101,7 @@
     {
         self.teebee.watchedEpisodesCount++;
         self.teebee.notWatchedEpisodesCount--;
-        self.teebee.seasons[StringInteger(self.teebeeEpisode.seasonNumber)] = @([self.teebee.seasons[StringInteger(self.teebeeEpisode.seasonNumber)] integerValue] + 1);
+        self.teebee.seasons[StringInteger(self.teebeeEpisode.seasonNumber)] = @([self.teebee.seasons[StringInteger((long)self.teebeeEpisode.seasonNumber)] integerValue] + 1);
         
         self.watchButton.selected = YES;
